blog-cms/requirements.txt
Flask==3.0.0
GitPython==3.1.40
python-frontmatter==1.0.0
python-dateutil==2.9.0
PyYAML==6.0.1
markdown2==2.4.13
Werkzeug==3.0.1
gunicorn==21.2.0
